genomeUtil
