+++ scripts/machinelearning/ai/Neat.lua
@@ -17,87 +17,6 @@ local MathUtil = require('util.MathUtil')
 local GenomeUtil = require('util.GenomeUtil')
 local MutationRate = require('machinelearning.ai.model.MutationRate')
 
--- Counts the number of non-matching innovation's in genes1 and genes2 then divides by
--- the number of genes of genes1 or genes2, whichever is greater
----@param genes1 Gene[]
----@param genes2 Gene[]
-local function disjoint(genes1, genes2)
-    local i1 = {}
-    local i2 = {}
-    local numberOfDisjointGenes = 0
-
-    for _, gene1 in pairs(genes1) do
-        i1[gene1.innovation] = true
-    end
-
-    for _, gene2 in pairs(genes2) do
-        i2[gene2.innovation] = true
-    end
-
-    -- for every gene2 that's not gene1, increment disjointGenes
-    for _, gene1 in pairs(genes1) do
-        if not i2[gene1.innovation] then
-            numberOfDisjointGenes = numberOfDisjointGenes + 1
-        end
-    end
-
-    -- for every gene2 that's not gene1, increment disjointGenes
-    for _, gene2 in pairs(genes2) do
-        if not i1[gene2.innovation] then
-            numberOfDisjointGenes = numberOfDisjointGenes + 1
-        end
-    end
-
-    local numberOfGenes = math.max(#genes1, #genes2)
-
-    return numberOfDisjointGenes / numberOfGenes
-end
-
--- Returns the average difference between genes1 and genes2 weights
----@param genes1 Gene[]
----@param genes2 Gene[]
-local function weights(genes1, genes2)
-    local i2 = {}
-    local sum = 0
-    local coincident = 0
-
-    for _, gene2 in pairs(genes2) do
-        i2[gene2.innovation] = gene2
-    end
-
-    for _, gene1 in pairs(genes1) do
-        if i2[gene1.innovation] ~= nil then
-            local gene2 = i2[gene1.innovation]
-            sum = sum + math.abs(gene1.weight - gene2.weight)
-            coincident = coincident + 1
-        end
-    end
-
-    return sum / coincident
-end
-
----@param genome1 Genome
----@param genome2 Genome
-local function isSameSpecies(genome1, genome2)
-    -- Number of matching genes divided by number of genes
-    local dd = Properties.deltaDisjoint * disjoint(genome1.genes, genome2.genes)
-    -- average difference between genes
-    local dw = Properties.deltaWeights * weights(genome1.genes, genome2.genes)
-
-    return dd + dw < Properties.deltaThreshold
-end
-
----@param pool Pool
-local function getTotalAverageFitnessRank(pool)
-    local totalRank = 0
-
-    for _, species in pairs(pool.species) do
-        totalRank = totalRank + species.averageFitnessRank
-    end
-
-    return totalRank
-end
-
 ---@param genome Genome
 local function pointMutate(genome, perturbChance)
     local step = genome.mutationRates.values.step
@@ -620,18 +539,7 @@ end
 
 ---@param pool Pool
 function Neat.removeWeakSpecies(pool)
-    ---@type Species[]
-    local survived = {}
-
-    local totalAverageFitnessRanks = getTotalAverageFitnessRank(pool)
-    for _, species in pairs(pool.species) do
-        local breed = math.floor((species.averageFitnessRank / totalAverageFitnessRanks) * pool:getNumberOfGenomes())
-        if breed >= 1 then
-            table.insert(survived, species)
-        end
-    end
-
-    pool.species = survived
+    pool.species = GenomeUtil.removeWeakSpecies(pool)
 end
 
 ---@param child Genome
@@ -640,7 +548,7 @@ function Neat:addToSpecies(child)
 
     for _, species in pairs(self.pool.species) do
         -- Because we are combining species that are the same, we just need to check the first one as they all match
-        if isSameSpecies(child, species.genomes[1]) then
+        if GenomeUtil.isSameSpecies(child, species.genomes[1]) then
             table.insert(species.genomes, child)
             speciesFound = true
             break
